Fixed imports and moved setup to runserver
