+++ Backend/Program.cs
@@ -1,18 +1,18 @@
 ﻿using Backend;
-using Backend.Utils;
+using Backend.Database;
 
 internal class Program
 {
     private static void Main(string[] args)
     {
-        TextMessageContext.SetUp();
-
         ServerStartupOptions serverOptions = ServerStartupOptions.GetServerStartupOptions();
         RunServer(serverOptions);
     }
 
     private static void RunServer(ServerStartupOptions serverOptions)
     {
+        TextMessageContext.SetUp();
+
         Console.WriteLine("Hello, World!");
         Console.WriteLine($"Server should start on port {serverOptions.Port}");
     }
